Fix the lint failure a3a1bea1 shipped, and retract the rmw overclaim
a3a1bea1 went up failing lint: the new odd-slot test multiplies two int
constants into a size_t array bound, which clang-tidy rejects as an implicit
widening. I ran `make test` and `make sanitize` after adding that test and not
`make lint` - verified the thing I had just written against two of the gates
and pushed it past the third. Cast to size_t.

Also replaced rmw_tickle.h's "overflow impossible by construction" with what was
actually measured, because it should not sit in a public header until the real
bound is known:

  - Held samples exceed the tracking window by a small margin - held_peak
    270-272 against 256, on loopback and on the HIL rig in every one of twelve
    reps. Likeliest a transient between the watermark advancing and the drain
    after it; not yet shown. So the bound is approximately the window, not the
    window.
  - The dramatic "failure" - 469,081 overflows at a wide window - was not the
    construction failing. It was a3a1bea1's memory bug reading neighbouring
    memory as slots. Fixed, the same run held 4095 of 4096 and overflowed zero
    times.

Both corrections go in the same direction as the day's: a claim stated as
following from construction, checked against one run that happened not to hit
the edge, and written into a header as though it had been proved.

+++ tests/test_reliable_pubsub.c
@@ -2462,7 +2462,7 @@ static void test_reliable_releases_held_samples_when_gap_is_abandoned(void) {
 // fails if core ever steps past the end again.
 #define ODD_SLOTS 4
 #define ODD_SLOT_BYTES ((uint16_t)(sizeof(struct tt_ReorderSlot) + 12 + 5)) // deliberately not a multiple of 8
-static uint64_t odd_storage[(ODD_SLOTS * ODD_SLOT_BYTES) / sizeof(uint64_t)];
+static uint64_t odd_storage[((size_t)ODD_SLOTS * ODD_SLOT_BYTES) / sizeof(uint64_t)];
 
 static void test_reorder_stays_inside_an_odd_sized_buffer(void) {
     test_mock_reset();
